--- a/src/core/renderers/webgl/WebGlCoreRenderer.ts
+++ b/src/core/renderers/webgl/WebGlCoreRenderer.ts
@@ -231,81 +231,49 @@ export class WebGlCoreRenderer extends CoreRenderer {
    */
   addQuad(params: QuadOptions) {
     const { fQuadBuffer, uiQuadBuffer } = this;
-    const {
-      width,
-      height,
-      colorTl,
-      colorTr,
-      colorBl,
-      colorBr,
-      textureOptions,
-      shader,
-      shaderProps,
-      alpha,
-      clippingRect,
-      tx,
-      ty,
-      ta,
-      tb,
-      tc,
-      td,
-      renderCoords,
-      rtt: renderToTexture,
-      parentHasRenderTexture,
-      framebufferDimensions,
-    } = params;
-    let { texture } = params;
+    let texture = params.texture || this.defaultTexture;
 
     /**
      * If the shader props contain any automatic properties, update it with the
      * current dimensions and or alpha that will be used to render the quad.
      */
-    if (shaderProps !== null) {
-      if (hasOwn(shaderProps, '$dimensions')) {
-        const dimensions = shaderProps.$dimensions as Dimensions;
-        dimensions.width = width;
-        dimensions.height = height;
+    if (params.shaderProps !== null) {
+      if (hasOwn(params.shaderProps, '$dimensions') == true) {
+        const dimensions = params.shaderProps.$dimensions as Dimensions;
+        dimensions.width = params.width;
+        dimensions.height = params.height;
       }
 
-      if (hasOwn(shaderProps, '$alpha')) {
-        shaderProps.$alpha = alpha;
+      if (hasOwn(params.shaderProps, '$alpha') === true) {
+        params.shaderProps.$alpha = params.alpha;
       }
     }
 
-    texture = texture ?? this.defaultTexture;
     assertTruthy(texture instanceof Texture, 'Invalid texture type');
 
     let { curBufferIdx: bufferIdx, curRenderOp } = this;
-    const targetDims = {
-      width,
-      height,
-    };
-    const targetShader = shader || this.defaultShader;
+    const targetDims = { width: -1, height: -1 };
+    targetDims.width = params.width;
+    targetDims.height = params.height;
+
+    const targetShader = params.shader || this.defaultShader;
     assertTruthy(targetShader instanceof WebGlCoreShader);
 
-    if (!this.reuseRenderOp(params)) {
+    if (this.reuseRenderOp(params) === false) {
       // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
       this.newRenderOp(
         targetShader,
-        shaderProps as Record<string, unknown>,
-        alpha,
+        params.shaderProps as Record<string, unknown>,
+        params.alpha,
         targetDims,
-        clippingRect,
+        params.clippingRect,
         bufferIdx,
-        renderToTexture,
-        parentHasRenderTexture,
-        framebufferDimensions,
+        params.rtt,
+        params.parentHasRenderTexture,
+        params.framebufferDimensions,
       );
       curRenderOp = this.curRenderOp;
       assertTruthy(curRenderOp);
-    }
-
-    const flipX = textureOptions?.flipX ?? false;
-    let flipY = textureOptions?.flipY ?? false;
-
-    // always flip flipY for render textures
-    if (texture instanceof RenderTexture) {
-      flipY = !flipY;
     }
 
     let texCoordX1 = 0;
@@ -324,37 +292,47 @@ export class WebGlCoreRenderer extends CoreRenderer {
       texture = texture.parentTexture;
     }
 
-    const resizeMode = textureOptions?.resizeMode ?? false;
-
-    if (texture instanceof ImageTexture) {
-      if (resizeMode && texture.dimensions) {
-        const { width: tw, height: th } = texture.dimensions;
-        if (resizeMode.type === 'cover') {
-          const scaleX = width / tw;
-          const scaleY = height / th;
-          const scale = Math.max(scaleX, scaleY);
-          const precision = 1 / scale;
-          // Determine based on width
-          if (scale && scaleX && scaleX < scale) {
-            const desiredSize = precision * width;
-            texCoordX1 = (1 - desiredSize / tw) * (resizeMode.clipX ?? 0.5);
-            texCoordX2 = texCoordX1 + desiredSize / tw;
-          }
-          // Determine based on height
-          if (scale && scaleY && scaleY < scale) {
-            const desiredSize = precision * height;
-            texCoordY1 = (1 - desiredSize / th) * (resizeMode.clipY ?? 0.5);
-            texCoordY2 = texCoordY1 + desiredSize / th;
-          }
+    if (
+      texture instanceof ImageTexture &&
+      params.textureOptions !== null &&
+      params.textureOptions.resizeMode !== undefined &&
+      texture.dimensions !== null
+    ) {
+      const resizeMode = params.textureOptions.resizeMode;
+      const { width: tw, height: th } = texture.dimensions;
+      if (resizeMode.type === 'cover') {
+        const scaleX = params.width / tw;
+        const scaleY = params.height / th;
+        const scale = Math.max(scaleX, scaleY);
+        const precision = 1 / scale;
+        // Determine based on width
+        if (scale && scaleX && scaleX < scale) {
+          const desiredSize = precision * params.width;
+          texCoordX1 = (1 - desiredSize / tw) * (resizeMode.clipX ?? 0.5);
+          texCoordX2 = texCoordX1 + desiredSize / tw;
+        }
+        // Determine based on height
+        if (scale && scaleY && scaleY < scale) {
+          const desiredSize = precision * params.height;
+          texCoordY1 = (1 - desiredSize / th) * (resizeMode.clipY ?? 0.5);
+          texCoordY2 = texCoordY1 + desiredSize / th;
         }
       }
     }
 
     // Flip texture coordinates if dictated by texture options
-    if (flipX) {
-      [texCoordX1, texCoordX2] = [texCoordX2, texCoordX1];
+    let flipY = 0;
+    if (params.textureOptions !== null) {
+      if (params.textureOptions.flipX === true) {
+        [texCoordX1, texCoordX2] = [texCoordX2, texCoordX1];
+      }
+
+      // convert to integer for bitwise operation below
+      flipY = +(params.textureOptions.flipY || false);
     }
-    if (flipY) {
+
+    // Eitherone should be true
+    if (flipY ^ +(texture instanceof RenderTexture)) {
       [texCoordY1, texCoordY2] = [texCoordY2, texCoordY1];
     }
 
@@ -364,99 +342,100 @@ export class WebGlCoreRenderer extends CoreRenderer {
 
     curRenderOp = this.curRenderOp;
     assertTruthy(curRenderOp);
-    if (renderCoords) {
-      const { x1, y1, x2, y2, x3, y3, x4, y4 } = renderCoords;
+    if (params.renderCoords) {
       // Upper-Left
-      fQuadBuffer[bufferIdx++] = x1; // vertexX
-      fQuadBuffer[bufferIdx++] = y1; // vertexY
+      fQuadBuffer[bufferIdx++] = params.renderCoords.x1; // vertexX
+      fQuadBuffer[bufferIdx++] = params.renderCoords.y1; // vertexY
       fQuadBuffer[bufferIdx++] = texCoordX1; // texCoordX
       fQuadBuffer[bufferIdx++] = texCoordY1; // texCoordY
-      uiQuadBuffer[bufferIdx++] = colorTl; // color
+      uiQuadBuffer[bufferIdx++] = params.colorTl; // color
       fQuadBuffer[bufferIdx++] = textureIdx; // texIndex
 
       // Upper-Right
-      fQuadBuffer[bufferIdx++] = x2;
-      fQuadBuffer[bufferIdx++] = y2;
+      fQuadBuffer[bufferIdx++] = params.renderCoords.x2;
+      fQuadBuffer[bufferIdx++] = params.renderCoords.y2;
       fQuadBuffer[bufferIdx++] = texCoordX2;
       fQuadBuffer[bufferIdx++] = texCoordY1;
-      uiQuadBuffer[bufferIdx++] = colorTr;
+      uiQuadBuffer[bufferIdx++] = params.colorTr;
       fQuadBuffer[bufferIdx++] = textureIdx;
 
       // Lower-Left
-      fQuadBuffer[bufferIdx++] = x4;
-      fQuadBuffer[bufferIdx++] = y4;
+      fQuadBuffer[bufferIdx++] = params.renderCoords.x4;
+      fQuadBuffer[bufferIdx++] = params.renderCoords.y4;
       fQuadBuffer[bufferIdx++] = texCoordX1;
       fQuadBuffer[bufferIdx++] = texCoordY2;
-      uiQuadBuffer[bufferIdx++] = colorBl;
+      uiQuadBuffer[bufferIdx++] = params.colorBl;
       fQuadBuffer[bufferIdx++] = textureIdx;
 
       // Lower-Right
-      fQuadBuffer[bufferIdx++] = x3;
-      fQuadBuffer[bufferIdx++] = y3;
+      fQuadBuffer[bufferIdx++] = params.renderCoords.x3;
+      fQuadBuffer[bufferIdx++] = params.renderCoords.y3;
       fQuadBuffer[bufferIdx++] = texCoordX2;
       fQuadBuffer[bufferIdx++] = texCoordY2;
-      uiQuadBuffer[bufferIdx++] = colorBr;
+      uiQuadBuffer[bufferIdx++] = params.colorBr;
       fQuadBuffer[bufferIdx++] = textureIdx;
-    } else if (tb !== 0 || tc !== 0) {
+    } else if (params.tb !== 0 || params.tc !== 0) {
       // Upper-Left
-      fQuadBuffer[bufferIdx++] = tx; // vertexX
-      fQuadBuffer[bufferIdx++] = ty; // vertexY
+      fQuadBuffer[bufferIdx++] = params.tx; // vertexX
+      fQuadBuffer[bufferIdx++] = params.ty; // vertexY
       fQuadBuffer[bufferIdx++] = texCoordX1; // texCoordX
       fQuadBuffer[bufferIdx++] = texCoordY1; // texCoordY
-      uiQuadBuffer[bufferIdx++] = colorTl; // color
+      uiQuadBuffer[bufferIdx++] = params.colorTl; // color
       fQuadBuffer[bufferIdx++] = textureIdx; // texIndex
 
       // Upper-Right
-      fQuadBuffer[bufferIdx++] = tx + width * ta;
-      fQuadBuffer[bufferIdx++] = ty + width * tc;
+      fQuadBuffer[bufferIdx++] = params.tx + params.width * params.ta;
+      fQuadBuffer[bufferIdx++] = params.ty + params.width * params.tc;
       fQuadBuffer[bufferIdx++] = texCoordX2;
       fQuadBuffer[bufferIdx++] = texCoordY1;
-      uiQuadBuffer[bufferIdx++] = colorTr;
+      uiQuadBuffer[bufferIdx++] = params.colorTr;
       fQuadBuffer[bufferIdx++] = textureIdx;
 
       // Lower-Left
-      fQuadBuffer[bufferIdx++] = tx + height * tb;
-      fQuadBuffer[bufferIdx++] = ty + height * td;
+      fQuadBuffer[bufferIdx++] = params.tx + params.height * params.tb;
+      fQuadBuffer[bufferIdx++] = params.ty + params.height * params.td;
       fQuadBuffer[bufferIdx++] = texCoordX1;
       fQuadBuffer[bufferIdx++] = texCoordY2;
-      uiQuadBuffer[bufferIdx++] = colorBl;
+      uiQuadBuffer[bufferIdx++] = params.colorBl;
       fQuadBuffer[bufferIdx++] = textureIdx;
 
       // Lower-Right
-      fQuadBuffer[bufferIdx++] = tx + width * ta + height * tb;
-      fQuadBuffer[bufferIdx++] = ty + width * tc + height * td;
+      fQuadBuffer[bufferIdx++] =
+        params.tx + params.width * params.ta + params.height * params.tb;
+      fQuadBuffer[bufferIdx++] =
+        params.ty + params.width * params.tc + params.height * params.td;
       fQuadBuffer[bufferIdx++] = texCoordX2;
       fQuadBuffer[bufferIdx++] = texCoordY2;
-      uiQuadBuffer[bufferIdx++] = colorBr;
+      uiQuadBuffer[bufferIdx++] = params.colorBr;
       fQuadBuffer[bufferIdx++] = textureIdx;
     } else {
       // Calculate the right corner of the quad
       // multiplied by the scale
-      const rightCornerX = tx + width * ta;
-      const rightCornerY = ty + height * td;
+      const rightCornerX = params.tx + params.width * params.ta;
+      const rightCornerY = params.ty + params.height * params.td;
 
       // Upper-Left
-      fQuadBuffer[bufferIdx++] = tx; // vertexX
-      fQuadBuffer[bufferIdx++] = ty; // vertexY
+      fQuadBuffer[bufferIdx++] = params.tx; // vertexX
+      fQuadBuffer[bufferIdx++] = params.ty; // vertexY
       fQuadBuffer[bufferIdx++] = texCoordX1; // texCoordX
       fQuadBuffer[bufferIdx++] = texCoordY1; // texCoordY
-      uiQuadBuffer[bufferIdx++] = colorTl; // color
+      uiQuadBuffer[bufferIdx++] = params.colorTl; // color
       fQuadBuffer[bufferIdx++] = textureIdx; // texIndex
 
       // Upper-Right
       fQuadBuffer[bufferIdx++] = rightCornerX;
-      fQuadBuffer[bufferIdx++] = ty;
+      fQuadBuffer[bufferIdx++] = params.ty;
       fQuadBuffer[bufferIdx++] = texCoordX2;
       fQuadBuffer[bufferIdx++] = texCoordY1;
-      uiQuadBuffer[bufferIdx++] = colorTr;
+      uiQuadBuffer[bufferIdx++] = params.colorTr;
       fQuadBuffer[bufferIdx++] = textureIdx;
 
       // Lower-Left
-      fQuadBuffer[bufferIdx++] = tx;
+      fQuadBuffer[bufferIdx++] = params.tx;
       fQuadBuffer[bufferIdx++] = rightCornerY;
       fQuadBuffer[bufferIdx++] = texCoordX1;
       fQuadBuffer[bufferIdx++] = texCoordY2;
-      uiQuadBuffer[bufferIdx++] = colorBl;
+      uiQuadBuffer[bufferIdx++] = params.colorBl;
       fQuadBuffer[bufferIdx++] = textureIdx;
 
       // Lower-Right
@@ -464,7 +443,7 @@ export class WebGlCoreRenderer extends CoreRenderer {
       fQuadBuffer[bufferIdx++] = rightCornerY;
       fQuadBuffer[bufferIdx++] = texCoordX2;
       fQuadBuffer[bufferIdx++] = texCoordY2;
-      uiQuadBuffer[bufferIdx++] = colorBr;
+      uiQuadBuffer[bufferIdx++] = params.colorBr;
       fQuadBuffer[bufferIdx++] = textureIdx;
     }
     // Update the length of the current render op
@@ -536,14 +515,12 @@ export class WebGlCoreRenderer extends CoreRenderer {
         throw new Error('Unable to add texture to render op');
       }
 
-      const { shader, shaderProps, dimensions, clippingRect, alpha } =
-        curRenderOp;
       this.newRenderOp(
-        shader,
-        shaderProps,
-        alpha,
-        dimensions,
-        clippingRect,
+        curRenderOp.shader,
+        curRenderOp.shaderProps,
+        curRenderOp.alpha,
+        curRenderOp.dimensions,
+        curRenderOp.clippingRect,
         bufferIdx,
       );
       return this.addTexture(texture, bufferIdx, true);
@@ -556,7 +533,7 @@ export class WebGlCoreRenderer extends CoreRenderer {
    * @param params
    * @returns
    */
-  reuseRenderOp(params: QuadOptions) {
+  reuseRenderOp(params: QuadOptions): boolean {
     const { shader, shaderProps, parentHasRenderTexture, rtt, clippingRect } =
       params;